refactor to standard and book modes
