+++ src/components/book/book.tsx
@@ -1,0 +1,12 @@
+import React from "react"
+
+
+const Book = () => {
+  return (
+    <section>
+        Book Mode
+    </section>
+  )
+}
+
+export default Book
