+++ src/styles/theme-selector.module.css
@@ -1,0 +1,3 @@
+.themeLabel {
+	padding-right: 10px;
+}
